+++ doc/whats-new.rst
@@ -190,6 +190,9 @@ Enhancements
   (:issue:`1485`).
   By `Joe Hamman <https://github.com/jhamman>`_.
 
+- New :py:meth:`~xarray.DataArray.rank`. Requires bottleneck (:issue:`1731`).
+  By `0x0L <https://github.com/0x0L>`_.
+
 **Performance improvements**
 
 - :py:func:`~xarray.concat` was computing variables that aren't in memory